requirements.txt
Flask==2.3.2
tensorflow==2.18.1
Pillow==11.3.0
numpy==1.26.4    
flask-cors==3.0.10